pyproject.toml
[project]
name = "mcp-server-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.1.0",
    "groq>=1.1.1",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.2",
    "trafilatura>=2.0.0",
]
